backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dateutil
pytest
python-dotenv
